Introduce extraLatexPackages argument
In this commit we are introducing a new argument to a mkDoc function called
extraLatexPackages. This is a attribute set holding additional latex packages
that are not included in the texlive build dependency by default.

This allows projects that are using mkDoc to add additional LaTeX packages that
they need, without polluting the danalib/mkDoc.

Because this argument has a natural default value (an empty attribute set),
we've modified the definition of mkDoc to take a single attribute set as an
argument. Therefore this is not completely backward-compatible change - the
projects that are using mkDoc will have to modify their function calls

Co-authored-by: Lukasz Golebiewski <lukasz.golebiewski@gmail.com>